LPS-53617 EditFeedAction: move logic to portlet class (no logic changes)
